requirements.txt
Flask>=2.0.0
Werkzeug>=2.0.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0

